requirements.txt
Faker==2.0.4
robotframework==3.1.2
robotframework-datetime-tz==1.0.6
robotframework-faker==4.2.0
robotframework-screencaplibrary==1.3.1
robotframework-seleniumlibrary==4.1.0
robotframeworklexer==1.1
selenium==3.141.0
text-unidecode==1.3
robotframework-pabot==0.98
robotframework-angularjs==0.0.10
robotframework-oracledb==2.0.2
JayDeBeApi==1.2.3

